feat(app): compact single-line chat composer + split assistant client
Empty chat opens as a full-width single-line composer that grows as you
type, reverting to the full composer once the first message is sent. Drop
the skills section (tabs + cards) from the chat empty state and the skills
prefetch from both chat pages. Allow drafting the next message while a
response streams (Enter ignored mid-stream; button stays a Stop control).
Split workspace-assistant-client into chat-composer, chat-message, and
message-part for maintainability.
