+++ apps/app/src/app/(app)/(pending-not-allowed)/[slug]/(workspace)/_components/chat-message.test.tsx
@@ -48,7 +48,7 @@ vi.mock("@repo/ui/components/ai-elements/tool", () => ({
   ToolOutput: () => <div>Tool output</div>,
 }));
 
-const { ChatMessage } = await import("./workspace-assistant-client");
+const { ChatMessage } = await import("./chat-message");
 
 function Harness({
   message,
